refactor: remove unused paste action context and streamline paste handling
